feat: funding rate precision i64 bps → i128 e9 (ppb) per spec v12.14.0
Phase 0+3 of v12.14.0 upgrade: funding rate precision change.

Engine changes:
- MAX_ABS_FUNDING_BPS_PER_SLOT (i64, 10_000) replaced by
  MAX_ABS_FUNDING_E9_PER_SLOT (i128, 1_000_000_000)
- funding_rate_bps_per_slot_last: i64 → funding_rate_e9_per_slot_last: i128
- All public methods: funding_rate: i64 → funding_rate_e9: i128
- accrue_market_to divisor: 10_000 → 1_000_000_000
- validate_funding_rate → validate_funding_rate_e9
- recompute_r_last_from_final_state: i64 → i128

Arithmetic fits: max fund_num = 10^12 * 10^9 * 65535 ≈ 6.55e25
(well within i128::MAX ≈ 1.7e38).

Test updates: all funding rates scaled from bps to ppb (e.g.,
500 bps = 50_000_000 ppb). Concrete test values updated to produce
nonzero fund_term with the larger divisor.

Version refs updated to v12.14.0.

+++ tests/proofs_lazy_ak.rs
@@ -53,7 +53,7 @@ fn t1_8_adl_deficit_only_lazy_equals_eager() {
 
     let eager_loss = ((q_base as i32) * (d as i32)) / (oi as i32);
 
-    let delta_k_abs = ((d as u32) * (a_side as u32) + (oi as u32) - 1) / (oi as u32);
+    let delta_k_abs = ((d as u16) * (a_side as u16) + (oi as u16) - 1) / (oi as u16);
     let delta_k = -(delta_k_abs as i32);
     let k_after = k_init + delta_k;
     let k_diff = k_after - k_init;
@@ -91,7 +91,7 @@ fn t1_9_adl_quantity_plus_deficit_lazy_conservative() {
     assert!(lazy_q <= eager_q, "lazy must not exceed eager quantity");
     assert!(eager_q - lazy_q <= 1, "lazy error bounded by 1 base unit");
 
-    let delta_k_abs = ((d as u32) * (a_old as u32) + (oi as u32) - 1) / (oi as u32);
+    let delta_k_abs = ((d as u16) * (a_old as u16) + (oi as u16) - 1) / (oi as u16);
     let delta_k = -(delta_k_abs as i32);
     let lazy_loss = -lazy_pnl(basis_q, delta_k, a_old);
     let eager_loss = ((q_base as i32) * (d as i32)) / (oi as i32);
@@ -124,7 +124,7 @@ fn t1_8b_adl_deficit_lazy_conservative_symbolic_a_basis() {
 
     let eager_loss = ((q_base as i32) * (d as i32)) / (oi as i32);
 
-    let delta_k_abs = ((d as u32) * (a_basis as u32) + (oi as u32) - 1) / (oi as u32);
+    let delta_k_abs = ((d as u16) * (a_basis as u16) + (oi as u16) - 1) / (oi as u16);
     let delta_k = -(delta_k_abs as i32);
     let lazy_loss_raw = lazy_pnl(basis_q, delta_k, a_basis);
 
@@ -463,7 +463,7 @@ fn t6_24_worked_example_regression() {
     let oi_post = oi - q_close;
     assert!(oi_post > 0);
 
-    let delta_k_abs = ((d as u32) * (a_long as u32) + (oi as u32) - 1) / (oi as u32);
+    let delta_k_abs = ((d as u16) * (a_long as u16) + (oi as u16) - 1) / (oi as u16);
     assert!(delta_k_abs == 64);
     let delta_k = -(delta_k_abs as i32);
     k_long = k_long + delta_k;
@@ -495,7 +495,7 @@ fn t6_25_pure_pnl_bankruptcy_regression() {
     let a_opp = S_ADL_ONE;
     let basis_q = (q_base as u16) * S_POS_SCALE;
 
-    let delta_k_abs = ((d as u32) * (a_opp as u32) + (oi as u32) - 1) / (oi as u32);
+    let delta_k_abs = ((d as u16) * (a_opp as u16) + (oi as u16) - 1) / (oi as u16);
     assert!(delta_k_abs > 0);
 
     let delta_k = -(delta_k_abs as i32);
